// src/app/interfaces/user-profile.interface.ts
export interface UserProfileInterface {
    fullname: string;
    userId: string;
    job_title: string;
    experience: string;
    age: string;
    current_sal_min: number;
    current_sal_max: number;
    expected_sal_min: number;
    expected_sal_max: number;
    education_levels: string;
    categories: string;
    description: string;
    photo_url: string;
    phone_number: string;
    email: string;
    website: string;
    address: string;
    country: string;
    city: string;
}